Initial refactoring to put all scheduled feeds into one Cloud Run service
