+++ cmd/go.mod
@@ -1,0 +1,7 @@
+module github.com/ossf/package-feeds
+
+go 1.15
+
+replace github.com/ossf/package-feeds/feeds/pypi => ../feeds/pypi
+
+require github.com/ossf/package-feeds/feeds/pypi v0.0.0-00010101000000-000000000000 // indirect
