--- feeds/crates/go.mod
@@ -1,8 +1,0 @@
-module github.com/ossf/package-feeds/feeds/crates
-
-go 1.15
-
-require (
-	github.com/jordan-wright/ossmalware v0.0.0-20201130035417-0b6db94c48ec
-	gocloud.dev v0.21.0
-)
