+++ feeds/go.mod
@@ -1,0 +1,5 @@
+module github.com/ossf/package-feeds/feeds
+
+go 1.15
+
+require github.com/sirupsen/logrus v1.7.0 // indirect
